Implement fuzzy matching to better find nodes.
If an exact match based on the Node hash cannot be found, that means one
of its parameters have changed. Try to first find a node based on the
id, and if that doesn't work and the id is unique within the module,
then use that.

Matching node ids would be something like "fp_text reference R12" or
"pad 1 smd" and failing that, unique short ids would be "path", "at",
or "tstamp", for examples.

Better matching will result in a more meaningful diff, as the nodes
appear in the same position and can be diff'ed as the same line.