Check that target file and .orig are identical
Running kicad-norm is supposed to be idempotent and it assumes that the
".orig" file, if it exists, is a copy of the target file.

However, it could be that the ".orig" file was left over from a previous
run and the target file has been edited since then. If so, this would
result in data loss after normalization. To prevent this, we check that
the files are identical first, before overwriting the target file.