Make findModule() use module "path" for matching
When swapping out component footprints for another, the Node ID will
change because it is composed of "module <footprint-id>". When that
happens, even though everything else like pin-to-net mapping and
reference values do not change, the diff engine may treat it as a total
change.

By modifying findModule() to associate these new "module" blocks to the
old ones, we can try to make sure its position within the file does not
change, and any child nodes which look similar to the old ones are also
mapped over with their sequence preserved. The "path" attribute in the
module looks to be unchanged, so that's what is being used for the
association here.