Remove .orig file after rewriting, by default
This eases the normal workflow: after commiting the rewritten file, the
.orig file will still be around. After more changes are made in KiCad,
and kicad-norm is re-run, it will now complain because the .orig file
and the new changes don't match up, and now the old .orig file needs to
be manually removed.

The downside of this change is that the original file saved by KiCad is
definitely removed, but hopefully our file hashing will ensure that the
rewritten file is semantically the same as the original, just
reshuffled.

You can still keep the orgiinal file around using -keep-orig flag.